Move the Jackson util code from package-private GitHubClient to public ResponseInfo.
