+++ src/main/java/org/kohsuke/github/GHEventInfo.java
@@ -172,7 +172,7 @@ public class GHEventInfo extends GitHubInteractiveObject {
      *             if payload cannot be parsed
      */
     public <T extends GHEventPayload> T getPayload(Class<T> type) throws IOException {
-        T v = GitHubClient.getMappingObjectReader(root()).readValue(payload.traverse(), type);
+        T v = ResponseInfo.getMappingObjectReader(root()).readValue(payload.traverse(), type);
         v.lateBind();
         return v;
     }
